Split ShareAction into a separate component for reusability
