+++ src/site/_includes/partials/subscribe-action.njk
@@ -1,0 +1,8 @@
+<a class="w-actions__fab w-actions__fab--subscribe gc-analytics-event"
+  data-category="web.dev"
+  data-label="subscribe, newsletter"
+  data-action="click"
+  href="/newsletter/"
+>
+  <span>Subscribe</span>
+</a>
