+++ src/site/content/en/live/index.njk
@@ -103,10 +103,15 @@ draft: true
         Share this event with your friends
       </p>
       <div class="w-event-buttons">
-        <a class="w-button w-button--primary w-button--with-icon gc-analytics-event" data-category="web.dev" data-label="live, share" data-action="click" data-icon="share" target="_blank"
-          href="#">
-          Share
-        </a>
+        <share-action class="gc-analytics-event"
+          data-category="web.dev"
+          data-label="live, share"
+          data-action="click"
+        >
+          <a class="w-button w-button--primary w-button--with-icon" data-icon="share">
+            Share
+          </a>
+        </share-action>
       </div>
     </div>
     <div class="w-event-section__column-right w-event-antihero">
