+++ src/site/content/en/tags/index.njk
@@ -12,7 +12,9 @@ pagination:
 {# It would be great to remove them as part of the v1 migration. #}
 <div class="tags-landing-page">
   {# Make sure Actions are first in the tab order. #}
-  <web-actions actions="subscribe"></web-actions>
+  <div class="w-actions">
+    {% include 'partials/subscribe-action.njk' %}
+  </div>
 
   <div class="w-layout-container w-pb--non">
     <header class="w-page-header">
