+++ src/styles/components/_actions.scss
@@ -22,6 +22,10 @@ $W_FAB_LARGE_RADIUS: 28px;
   z-index: 100;
 }
 
+.w-actions > * {
+  margin-top: 20px;
+}
+
 .w-actions__fab {
   align-items: center;
   background-color: $WEB_PRIMARY_COLOR;
@@ -106,10 +110,6 @@ $W_FAB_LARGE_RADIUS: 28px;
 
 .w-actions__fab:active::after {
   background-color: rgba($WHITE, .32);
-}
-
-.w-actions__fab + .w-actions__fab {
-  margin-top: 20px;
 }
 
 // FAB variations
